backend/requirements.txt
fastapi
pydantic
openai-agents
python-dotenv
